Update report.Rmd
Sorry for the flurry of reviewed documents today, and for not getting feedback to you earlier @Dewi-Amaliah! This is some comments on your methodology section. 

I think overall it's very good. I've suggested some changes to the section where you describe the literature review methodology - specifically splitting it into the abstract reviewing phase and the full paper reviewing phase. This will make it easier for you to describe what you did and which criteria you are referring to. 
I've also suggested some rewordings (these are truly suggestions, this is your own work, choose the words that you think is best). 

Other than that I've made a few suggestions in text (prefaced with LK to make it easy to find them). The main suggestion I have is to describe why we chose to do things/collect the information we did. The choices we made were very intentional, and it's good to demonstrate that you know the reasons! :) 

Let me know or tag me if you need clarification or would like to discuss further.